+++ R/plot_auxindex.R
@@ -1,7 +1,8 @@
 # Residual vs Index Plot.
 
 # Creates a residual plot with the input residuals and predicted values
-plot_auxindex <- function(resid, smoother, theme, axis.text.size, title.text.size, title.opt){
+plot_auxindex <- function(resid, smoother, theme, axis.text.size, title.text.size, title.opt,
+                          alpha){
 
   ## Creation of Values to Plot -----------------------------------------------------
 
@@ -14,15 +15,25 @@ plot_auxindex <- function(resid, smoother, theme, axis.text.size, title.text.siz
   ## Creation of Plot ---------------------------------------------------------------
 
   # Create the residual plot
-  plot <- ggplot(data = model_values, aes_string(x = "Observation", y = "Residual")) +
-    geom_point() +
+  plot <- 
+    ggplot(
+      data = model_values,
+      mapping = aes(x = {Observation}, y = {Residual})
+    ) +
+    geom_point(alpha = alpha) +
     geom_abline(slope = 0, intercept = 0, color = "blue") +
     labs(x = "Observation Number", y = "Residuals")
 
   # If smoother is set to true, add it to the plot
   if (smoother == TRUE){
     plot <- plot +
-      geom_smooth(method = "loess", se = FALSE, color = "red", size = 0.5)
+      geom_smooth(
+        method = "loess", 
+        se = FALSE, 
+        color = "red", 
+        linewidth = 0.5,
+        formula = 'y ~ x'
+      )
   }
 
   # Add theme to plot
@@ -34,7 +45,7 @@ plot_auxindex <- function(resid, smoother, theme, axis.text.size, title.text.siz
     plot <- plot + theme_grey()
   }
 
-  # Set text size of title and axis lables, determine whether to include a title,
+  # Set text size of title and axis labels, determine whether to include a title,
   # and return plot
   if(title.opt == TRUE){
     plot +