+++ R/plot_constlev.R
@@ -1,7 +1,8 @@
 # Residual-Leverage plot.
 
 # Creates a plot of the residuals versus leverage from a model
-plot_constlev <- function(model, type, theme, axis.text.size, title.text.size, title.opt){
+plot_constlev <- function(model, type, theme, axis.text.size, title.text.size, title.opt,
+                          alpha = alpha){
 
   ## Creation of Values to Plot -----------------------------------------------------
 
@@ -46,9 +47,9 @@ plot_constlev <- function(model, type, theme, axis.text.size, title.text.size, t
     }
   }
 
-  # Compute the values for the lowess curve
-  model_values$Lowess.x <- lowess(x = model_values$Variables, y = model_values$Std_Res)$x
-  model_values$Lowess.y <- lowess(x = model_values$Variables, y = model_values$Std_Res)$y
+  # Compute the values for the lowess curve #Removed 
+  #model_values$Lowess.x <- lowess(x = model_values$Variables, y = model_values$Std_Res)$x
+  #model_values$Lowess.y <- lowess(x = model_values$Variables, y = model_values$Std_Res)$y
 
   ## Creation of Labels -------------------------------------------------------------
 
@@ -70,11 +71,16 @@ plot_constlev <- function(model, type, theme, axis.text.size, title.text.size, t
   ## Creation of Plot ---------------------------------------------------------------
   # Create the constant leverage plot
 
-  plot <- ggplot(data = model_values, aes_string(x = "Variables", y = "Std_Res"), na.rm=TRUE) +
-    geom_point(aes_string(group = "Data")) +
-    geom_line(aes_string(x = "Lowess.x", y = "Lowess.y"), color = "red", size = 0.5)+
-    geom_abline(slope = 0, intercept = 0, color = "blue", size = 0.5)+
-    xlab("Factor Level Combinations")+
+  plot <- 
+    ggplot(
+      data = model_values, 
+      mapping = aes(x = {Variables}, y = {Std_Res}), 
+      na.rm = TRUE
+    ) +
+    geom_point(aes(group = {Data}), alpha = alpha) +
+    #geom_line(aes(x = {Lowess.x}, y = {Lowess.y}), color = "red", linewidth = 0.5) +
+    geom_abline(slope = 0, intercept = 0, color = "blue", linewidth = 0.5) +
+    xlab("Factor Level Combinations") +
     ylab("Standardized Residuals")
 
   # Add theme to plot
@@ -86,7 +92,7 @@ plot_constlev <- function(model, type, theme, axis.text.size, title.text.size, t
     plot <- plot + theme_grey()
   }
 
-  # Set text size of title and axis lables, determine whether to include a title,
+  # Set text size of title and axis labels, determine whether to include a title,
   # and return plot
   if(title.opt == TRUE){
     plot +