+++ R/plot_ls.R
@@ -2,7 +2,7 @@
 
 # Creates a location-scale plot with the square root of the standardized residuals
 # versus predicted values from a model
-plot_ls <- function(model, type, smoother, theme, axis.text.size, title.text.size, title.opt){
+plot_ls <- function(model, type, smoother, theme, axis.text.size, title.text.size, title.opt, alpha){
 
   ## Creation of Values to Plot -----------------------------------------------------
 
@@ -40,40 +40,63 @@ plot_ls <- function(model, type, smoother, theme, axis.text.size, title.text.siz
   if (class(model)[1] == "lm"){
 
     # Location-scale plot for lm model
-    plot <- ggplot(data = model_values,
-                   mapping = aes_string(x = "Prediction", y = "Sqrt_Std_Res", label = "Data")) +
-      geom_point() +
-      labs(x = "Predicted Values", y = expression(sqrt(abs(" Standardized Residuals ")))) +
+    plot <- 
+      ggplot(
+        data = model_values,
+        mapping = aes(x = {Prediction}, y = {Sqrt_Std_Res}, label = {Data})
+      ) +
+      geom_point(alpha = alpha) +
+      labs(
+        x = "Predicted Values", 
+        y = expression(sqrt(abs(" Standardized Residuals ")))
+      ) +
       expand_limits(y = 0)
 
-  } else if (class(model)[1] == "glm"){
+  } else if (class(model)[1] == "glm") {
 
     # Location-scale plot for glm model with deviance residuals
-    if(is.na(type) | type == "response" | type == "deviance" | type == "stand.deviance"){
-      plot <- ggplot(data = model_values,
-                     mapping = aes_string(x = "Prediction", y = "Sqrt_Std_Res", label = "Data")) +
-        geom_point() +
-        labs(x = "Predicted Values",
-             y = expression(sqrt(abs(" Standardized Deviance Residuals ")))) +
+    if (is.na(type) | type == "response" | type == "deviance" | type == "stand.deviance"){
+      plot <- 
+        ggplot(
+          data = model_values,
+          mapping = aes(x = {Prediction}, y = {Sqrt_Std_Res}, label = {Data})
+        ) +
+        geom_point(alpha = alpha) +
+        labs(
+          x = "Predicted Values",
+          y = expression(sqrt(abs(" Standardized Deviance Residuals ")))
+        ) +
         expand_limits(y = 0)
 
     # Location-scale plot for glm model with Pearson residuals
-    } else if(type == "pearson" | type == "stand.pearson"){
-      plot <- ggplot(data = model_values,
-                     mapping = aes_string(x = "Prediction", y = "Sqrt_Std_Res", label = "Data")) +
-        geom_point() +
-        labs(x = "Predicted Values",
-             y = expression(sqrt(abs(" Standardized Pearson Residuals ")))) +
+    } else if (type == "pearson" | type == "stand.pearson") {
+      
+      plot <- 
+        ggplot(
+          data = model_values,
+          mapping = aes(x = {Prediction}, y = {Sqrt_Std_Res}, label = {Data})
+        ) +
+        geom_point(alpha = alpha) +
+        labs(
+          x = "Predicted Values",
+          y = expression(sqrt(abs(" Standardized Pearson Residuals ")))
+        ) +
         expand_limits(y = 0)
-
+      
     }
 
   }
 
-  # If smoother is set to true, add it to the plot
+  # If smoother is set to true, do not add it to the plot
   if (smoother == TRUE){
-    plot <- plot +
-      geom_smooth(method = "loess", se = FALSE, color = "red", size = 0.5)
+    plot <- plot #+
+      # geom_smooth(
+      #   method = "loess", 
+      #   se = FALSE, 
+      #   color = "red", 
+      #   linewidth = 0.5, 
+      #   formula = 'y ~ x'
+      # )
   }
 
   # Add theme to plot
@@ -85,7 +108,7 @@ plot_ls <- function(model, type, smoother, theme, axis.text.size, title.text.siz
     plot <- plot + theme_grey()
   }
 
-  # Set text size of title and axis lables, determine whether to include a title,
+  # Set text size of title and axis labels, determine whether to include a title,
   # and return plot
   if(title.opt == TRUE){
     plot +