+++ R/plot_yvp.R
@@ -1,7 +1,7 @@
 # Response versus predicted plot.
 
 # Creates a plot of the response variable versus the predicted values
-plot_yvp <- function(model, theme, axis.text.size, title.text.size, title.opt){
+plot_yvp <- function(model, theme, axis.text.size, title.text.size, title.opt, alpha){
 
   ## Creation of Values to Plot -----------------------------------------------------
 
@@ -49,10 +49,14 @@ plot_yvp <- function(model, theme, axis.text.size, title.text.size, title.opt){
 
 
   # Create the plot of response variable versus predicted values
-  plot <- ggplot(data = model_values,
-                 mapping = aes_string(x = "Predicted", y = "Response", label = "Data")) +
-    geom_point() +
+  plot <- 
+    ggplot(
+      data = model_values,
+      mapping = aes(x = {Predicted}, y = {Response}, label = {Data})
+    ) +
+    geom_point(alpha = alpha) +
     geom_abline(slope = 1, intercept = 0, color = "blue") +
+    coord_fixed() + 
     labs(x = "Predicted Values", y = y_label)
 
   # Add theme to plot
@@ -64,7 +68,7 @@ plot_yvp <- function(model, theme, axis.text.size, title.text.size, title.opt){
     plot <- plot + theme_grey()
   }
 
-  # Set text size of title and axis lables, determine whether to include a title,
+  # Set text size of title and axis labels, determine whether to include a title,
   # and return plot
   if(title.opt == TRUE){
     plot +